requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pdfplumber
python-docx
pandas
pdf2image
pytesseract
